% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPrescription.R
\name{LoadPrescription}
\alias{LoadPrescription}
\title{Load in Prescriptions}
\usage{
LoadPrescription(
  project_name,
  data_base_name,
  prescription = NULL,
  person = NULL,
  ATC_4th = FALSE
)
}
\arguments{
\item{project_name}{string}

\item{data_base_name}{string}

\item{prescription}{integer}

\item{person}{vector}

\item{ATC_4th}{boolean}
}
\value{

}
\description{
returns information based on the prescription filtered or a list of
ids and returns all the people who were given that prescription filtered
by, or all the people in the person ids list. There is also an option
depending whether you want to filter by the 4th or 5th level of the ATC
drug concept, as a broader concept classification may be wanted
(e.g. Glucocorticoids vs Salmeterol).
}
\examples{
LoadPrescription("yhcr-prd-phm-bia-core", "CY_CDM_V1_50k_Random",
                  prescription = 21603256, person = NULL,  ATC_4th = FALSE)
}
