% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryOmop.R
\name{QueryOmop}
\alias{QueryOmop}
\title{Perform Query}
\usage{
QueryOmop(project_name, sql)
}
\arguments{
\item{project_name}{string}

\item{sql}{string}
}
\value{
data frame
}
\description{
Carries out Query on the OMOP CDM
}
\examples{
project_name <- "yhcr-prd-phm-bia-core"
sql <- "SELECT *,
        cast(CONCAT(year_of_birth,'-',month_of_birth,'-',day_of_birth) as string)
        as Birthday FROM yhcr-prd-phm-bia-core.CY_CDM_V1_50k_Random.person
        LIMIT 1000"
person <- QueryOmop(project_name, sql)
}
